Refactored all the Epub Translating logic to its own cpp leaving main.cpp to just be start up code
